+++ CHANGELOG.md
@@ -10,6 +10,11 @@ speaking, considered to be the player's experience.
 
 ## [Unreleased]
 
+### Added
+
+- Added scroll up and scroll down prompts to the What's New screen.
+- Added scroll hints to the What's New screen.
+
 
 ## [3.1.9] — 2025-06-25